requirements.txt
numpy
pandas
matplotlib
pillow
torch
torchvision
scikit-learn
spacy
ipykernel